Move DashboardHome component to its own file
