+++ src/components/dashboard.tsx
@@ -2,11 +2,6 @@ import { HeadingLevels } from "@/types/text";
 import { Heading } from "./ui/heading";
 import { HTMLAttributes } from "react";
 import { Section } from "./ui/section";
-import { AppData } from "@/data/records/apps";
-import { DashboardCard } from "./dashboard-card";
-import { AppIcon, AppIconShapes } from "./ui/app-icon";
-import { Apps } from "@/types/apps";
-import Link from "next/link";
 import { VariantProps, cva } from "class-variance-authority";
 import { cn } from "@/utils/helpers";
 
@@ -36,7 +31,7 @@ const DashboardRoot = ({ children }: DashboardRootProps) => {
 
 interface DashboardTitleProps extends HTMLAttributes<HTMLHeadingElement> {}
 
-const DashboardTitle = ({ children }: DashboardTitleProps) => {
+export const DashboardTitle = ({ children }: DashboardTitleProps) => {
   return (
     <Heading
       level={HeadingLevels.h1}
@@ -87,7 +82,7 @@ interface DashboardContentShelfProps {
   children: React.ReactNode;
 }
 
-const DashboardContentShelf = ({
+export const DashboardContentShelf = ({
   headingText,
   shelfBehavior,
   children,
@@ -97,50 +92,6 @@ const DashboardContentShelf = ({
       <DashboardHeading>{headingText}</DashboardHeading>
       <DashboardShelf mobileBehavior={shelfBehavior}>{children}</DashboardShelf>
     </section>
-  );
-};
-
-export const DashboardHome = () => {
-  const appCards = Object.values(AppData).map((app) => {
-    return (
-      <Link href={app.href}>
-        <DashboardCard
-          img={
-            <AppIcon app={app.appType as Apps} shape={AppIconShapes.Square} />
-          }
-          title={app.name}
-          description={app.description}
-          small
-        />
-      </Link>
-    );
-  });
-
-  const dummyCard = (
-    <DashboardCard
-      img={<div className="h-full w-full bg-gray"></div>}
-      title={"Playlist"}
-      description={
-        "Lorem ipsum dolor sit amet consectetur, adipisicing elit. Sit excepturi earum voluptatibus quos laboriosam obcaecati recusandae quo unde modi quam nobis provident perspiciatis, a natus debitis omnis! Incidunt, quae cum."
-      }
-    />
-  );
-
-  const dummyCards = [];
-  for (let i = 0; i < 10; i++) {
-    dummyCards.push(dummyCard);
-  }
-
-  return (
-    <>
-      <DashboardTitle>Ready to mix?</DashboardTitle>
-      <DashboardContentShelf headingText="Start mixing" shelfBehavior="two-col">
-        {appCards}
-      </DashboardContentShelf>
-      <DashboardContentShelf headingText="Mix your playlists">
-        {dummyCards}
-      </DashboardContentShelf>
-    </>
   );
 };
